# textToPoints perf improvement