--- a/package.json
+++ b/package.json
@@ -22,6 +22,7 @@
   },
   "version": "1.9.4",
   "dependencies": {
+    "@davepagurek/bezier-path": "^0.0.2",
     "acorn": "^8.12.1",
     "acorn-walk": "^8.3.4",
     "colorjs.io": "^0.5.2",
@@ -81,9 +82,7 @@
   },
   "author": "",
   "husky": {
-    "hooks": {
-      
-    }
+    "hooks": {}
   },
   "msw": {
     "workerDirectory": [