… test/unit/visual/screenshots/Typography/textToContours/Fonts can be converted to points grouped by contour/000.png
… test/unit/visual/screenshots/Typography/textToPoints/Fonts can be converted to points/000.png
… test/unit/visual/screenshots/Typography/textToPoints/Sampling density can be changed/000.png